refactor(test): service 패키지명 오타 수정
